refactor(query): Remove redundant components
In favour of option to add data to hs-panel-container, there is no need for additional components.

Moved clear layer under feature widgets, as it clears all features of the layer, not the layer it self

Kept HsLayerWidgetContainerService for future enhancements.
