+++ projects/hslayers/src/components/query/layer-widgets/layer-widget-container.service.ts
@@ -1,6 +1,6 @@
 import {Injectable} from '@angular/core';
 
-import {HsPanelContainerService} from '../../../layout/panels/panel-container.service';
+import {HsPanelContainerService} from '../../layout/panels/panel-container.service';
 
 @Injectable({
   providedIn: 'root',
