+++ projects/hslayers/src/components/query/public-api.ts
@@ -14,9 +14,7 @@ export * from './query-popup.service';
 export * from './query-popup.service.model';
 export * from './query-popup-base.service';
 export * from './query-popup-widget-base.component';
-export * from './query-popup-feature/query-popup-feature.component';
-export * from './query-popup-feature/feature-widgets/feature-info.component';
-export * from './query-popup-feature/feature-widgets/feature-widget-container.service';
-export * from './query-popup-layer/query-popup-layer.component';
-export * from './query-popup-layer/layer-widgets/clear-layer.component';
-export * from './query-popup-layer/layer-widgets/layer-widget-container.service';
+export * from './feature-widgets/feature-info.component';
+export * from './feature-widgets/feature-widget-container.service';
+export * from './feature-widgets/clear-layer.component';
+export * from './layer-widgets/layer-widget-container.service';
