--- projects/hslayers/src/components/query/query-popup-feature/query-popup-feature.component.html
@@ -1,4 +1,0 @@
-<form>
-  <hs-panel-container [service]="hsFeatureWidgetContainerService" [data]="{feature: feature, attributesForHover: attributesForHover, service: service}">
-  </hs-panel-container>
-</form>
