--- projects/hslayers/src/components/query/query-popup-layer/layer-widgets/clear-layer.component.html
@@ -1,5 +1,0 @@
-<button class="btn btn-danger btn-block btn-sm" (click)="clearLayer(layerDesc.layer)"
-  *ngIf="hsLayerUtilsService.isLayerEditable(layerDesc.layer)" data-toggle="tooltip"
-  [title]="'QUERY.featurePopup.clearLayer'| translate">
-  {{'QUERY.featurePopup.clearLayer' | translate}}
-</button>
