--- projects/hslayers/src/components/query/query-popup-layer/query-popup-layer.component.html
@@ -1,4 +1,0 @@
-<form class="d-flex justify-content-center">
-  <hs-panel-container [service]="hsLayerWidgetContainerService" [data]="{layerDesc: this.layerDescriptor, service: this.service}">
-  </hs-panel-container>
-</form>
