+++ projects/hslayers/src/components/query/query-popup/query-popup.component.ts
@@ -11,8 +11,10 @@ import {Subscription} from 'rxjs';
 import {HsDialogComponent} from '../../layout/dialogs/dialog-component.interface';
 import {HsDialogItem} from '../../layout/dialogs/dialog-item';
 import {HsEventBusService} from '../../core/event-bus.service';
+import {HsFeatureWidgetContainerService} from '../feature-widgets/feature-widget-container.service';
 import {HsLanguageService} from '../../language/language.service';
 import {HsLayerUtilsService} from '../../utils/layer-utils.service';
+import {HsLayerWidgetContainerService} from '../layer-widgets/layer-widget-container.service';
 import {HsMapService} from '../../map/map.service';
 import {HsQueryPopupServiceModel} from '../query-popup.service.model';
 import {getFeatures} from '../../../common/feature-extensions';
@@ -23,8 +25,7 @@ import {getPopUp} from '../../../common/layer-extensions';
   templateUrl: './query-popup.component.html',
 })
 export class HsQueryPopupComponent
-  implements OnDestroy, HsDialogComponent, AfterViewInit
-{
+  implements OnDestroy, HsDialogComponent, AfterViewInit {
   getFeatures = getFeatures;
   olMapLoadsSubscription: Subscription;
   attributesForHover = [];
@@ -39,7 +40,9 @@ export class HsQueryPopupComponent
     public hsLanguageService: HsLanguageService,
     public hsLayerUtilsService: HsLayerUtilsService, //Used in template
     public hsMapService: HsMapService,
-    private ElementRef: ElementRef
+    private ElementRef: ElementRef,
+    public hsLayerWidgetContainerService: HsLayerWidgetContainerService,
+    public hsFeatureWidgetContainerService: HsFeatureWidgetContainerService
   ) {
     this.olMapLoadsSubscription = this.hsEventBusService.olMapLoads.subscribe(
       (map) => {
